+++ client/shared/src/components/CtaAlert.story.tsx
@@ -2,7 +2,9 @@ import React from 'react'
 
 import { DecoratorFn, Meta, Story } from '@storybook/react'
 
+// eslint-disable-next-line no-restricted-imports
 import { ExtensionRadialGradientIcon } from '@sourcegraph/web/src/components/CtaIcons'
+// eslint-disable-next-line no-restricted-imports
 import { WebStory } from '@sourcegraph/web/src/components/WebStory'
 
 import { CtaAlert } from './CtaAlert'