+++ client/wildcard/src/components/index.ts
@@ -1,16 +1,26 @@
 /** Component exports */
-export { Button, ButtonGroup } from './Button'
+export { Button, ButtonGroup, BUTTON_SIZES } from './Button'
 export type { ButtonGroupProps } from './Button'
 export { Alert, AlertLink } from './Alert'
 export { Container } from './Container'
-export { Checkbox, FlexTextArea, Input, RadioButton, Select, MultiSelect, TextArea, InputStatus } from './Form'
+export {
+    Checkbox,
+    FlexTextArea,
+    Input,
+    RadioButton,
+    Select,
+    MultiSelect,
+    TextArea,
+    InputStatus,
+    getInputStatus,
+} from './Form'
 export { Grid } from './Grid'
 export { LoadingSpinner } from './LoadingSpinner'
 export { PageHeader } from './PageHeader'
 export { PageSelector } from './PageSelector'
 export { Tabs, Tab, TabList, TabPanel, TabPanels, useTabsContext } from './Tabs'
 export { SourcegraphIcon } from './SourcegraphIcon'
-export { Badge, ProductStatusBadge } from './Badge'
+export { Badge, ProductStatusBadge, BADGE_VARIANTS, PRODUCT_STATUSES } from './Badge'
 export { Panel } from './Panel'
 export { Tooltip, TooltipController } from './Tooltip'
 export { Card, CardBody, CardHeader, CardList, CardSubtitle, CardText, CardTitle, CardFooter } from './Card'
@@ -39,7 +49,7 @@ export { Collapse, CollapseHeader, CollapsePanel } from './Collapse'
  */
 export type { FeedbackPromptSubmitEventHandler } from './Feedback'
 export type { AlertProps, AlertLinkProps } from './Alert'
-export type { MultiSelectProps, MultiSelectOption, MultiSelectState, SelectProps } from './Form'
+export type { MultiSelectProps, MultiSelectOption, MultiSelectState, SelectProps, InputProps } from './Form'
 export type { ButtonProps } from './Button'
 export type { ButtonLinkProps } from './ButtonLink'
 export type { LinkProps } from './Link'
@@ -47,3 +57,5 @@ export type { PopoverOpenEvent } from './Popover'
 export type { TabsProps } from './Tabs'
 export type { IconProps } from './Icon'
 export type { Point } from './Popover'
+export type { BadgeProps, ProductStatusType, BaseProductStatusBadgeProps } from './Badge'
+export type { AnchorLinkProps } from './Link'