+++ internal/gitserver/commands.go
@@ -1,0 +1,121 @@
+package gitserver
+
+import (
+	"bytes"
+	"context"
+	"net/mail"
+	"strconv"
+	"strings"
+
+	"github.com/sourcegraph/sourcegraph/internal/api"
+	"github.com/sourcegraph/sourcegraph/internal/gitserver/gitdomain"
+	"github.com/sourcegraph/sourcegraph/internal/lazyregexp"
+	"github.com/sourcegraph/sourcegraph/internal/trace/ot"
+	"github.com/sourcegraph/sourcegraph/lib/errors"
+)
+
+// ShortLogOptions contains options for (Repository).ShortLog.
+type ShortLogOptions struct {
+	Range string // the range for which stats will be fetched
+	After string // the date after which to collect commits
+	Path  string // compute stats for commits that touch this path
+}
+
+func (c *ClientImplementor) ShortLog(ctx context.Context, repo api.RepoName, opt ShortLogOptions) ([]*gitdomain.PersonCount, error) {
+	span, ctx := ot.StartSpanFromContext(ctx, "Git: ShortLog")
+	span.SetTag("Opt", opt)
+	defer span.Finish()
+
+	if opt.Range == "" {
+		opt.Range = "HEAD"
+	}
+	if err := checkSpecArgSafety(opt.Range); err != nil {
+		return nil, err
+	}
+
+	// We split the individual args for the shortlog command instead of -sne for easier arg checking in the allowlist.
+	args := []string{"shortlog", "-s", "-n", "-e", "--no-merges"}
+	if opt.After != "" {
+		args = append(args, "--after="+opt.After)
+	}
+	args = append(args, opt.Range, "--")
+	if opt.Path != "" {
+		args = append(args, opt.Path)
+	}
+	cmd := c.Command("git", args...)
+	cmd.Repo = repo
+	out, err := cmd.Output(ctx)
+	if err != nil {
+		return nil, errors.Errorf("exec `git shortlog -s -n -e` failed: %v", err)
+	}
+	return parseShortLog(out)
+}
+
+// logEntryPattern is the regexp pattern that matches entries in the output of the `git shortlog
+// -sne` command.
+var logEntryPattern = lazyregexp.New(`^\s*([0-9]+)\s+(.*)$`)
+
+func parseShortLog(out []byte) ([]*gitdomain.PersonCount, error) {
+	out = bytes.TrimSpace(out)
+	if len(out) == 0 {
+		return nil, nil
+	}
+	lines := bytes.Split(out, []byte{'\n'})
+	results := make([]*gitdomain.PersonCount, len(lines))
+	for i, line := range lines {
+		// example line: "1125\tJane Doe <jane@sourcegraph.com>"
+		match := logEntryPattern.FindSubmatch(line)
+		if match == nil {
+			return nil, errors.Errorf("invalid git shortlog line: %q", line)
+		}
+		// example match: ["1125\tJane Doe <jane@sourcegraph.com>" "1125" "Jane Doe <jane@sourcegraph.com>"]
+		count, err := strconv.Atoi(string(match[1]))
+		if err != nil {
+			return nil, err
+		}
+		addr, err := lenientParseAddress(string(match[2]))
+		if err != nil || addr == nil {
+			addr = &mail.Address{Name: string(match[2])}
+		}
+		results[i] = &gitdomain.PersonCount{
+			Count: int32(count),
+			Name:  addr.Name,
+			Email: addr.Address,
+		}
+	}
+	return results, nil
+}
+
+// lenientParseAddress is just like mail.ParseAddress, except that it treats
+// the following somewhat-common malformed syntax where a user has misconfigured
+// their email address as their name:
+//
+// 	foo@gmail.com <foo@gmail.com>
+//
+// As a valid name, whereas mail.ParseAddress would return an error:
+//
+// 	mail: expected single address, got "<foo@gmail.com>"
+//
+func lenientParseAddress(address string) (*mail.Address, error) {
+	addr, err := mail.ParseAddress(address)
+	if err != nil && strings.Contains(err.Error(), "expected single address") {
+		p := strings.LastIndex(address, "<")
+		if p == -1 {
+			return addr, err
+		}
+		return &mail.Address{
+			Name:    strings.TrimSpace(address[:p]),
+			Address: strings.Trim(address[p:], " <>"),
+		}, nil
+	}
+	return addr, err
+}
+
+// checkSpecArgSafety returns a non-nil err if spec begins with a "-", which
+// could cause it to be interpreted as a git command line argument.
+func checkSpecArgSafety(spec string) error {
+	if strings.HasPrefix(spec, "-") {
+		return errors.Errorf("invalid git revision spec %q (begins with '-')", spec)
+	}
+	return nil
+}